feat(schemas): Validate all source topics have schemas defined
There's a few temporary exclusions for schemas which still need to be created.
